Refactor: Update import for bff method
